cleanup controllers
