+++ mobile/app/wallet/assets/ethereum/config/abi/general-abi.ts
